Day 7, Part 2: Update return type to Long and fix test assertions

+++ src/main/kotlin/day07/Day07.kt
@@ -21,7 +21,7 @@ class Day07 {
         return result
     }
 
-    fun part2(input: List<CharArray>): Int {
-        return 1
+    fun part2(input: List<CharArray>): Long {
+        return 1L
     }
 }
